BREAKING_CHANGE: remove all security stuff
