--- xpresso/binders/_security/oauth2.py
@@ -1,174 +1,0 @@
-import sys
-from typing import Any, Dict, List, Optional, Union
-
-if sys.version_info < (3, 8):
-    from typing_extensions import Literal
-else:
-    from typing import Literal
-
-from pydantic import BaseModel
-from starlette.requests import Request
-from starlette.status import HTTP_401_UNAUTHORIZED
-
-from xpresso.binders._param_functions import FormEncodedField
-from xpresso.binders._security.utils import get_authorization_scheme_param
-from xpresso.binders.api import SecurityBase
-from xpresso.exceptions import HTTPException
-from xpresso.openapi.models import OAuth2 as OAuth2Model
-from xpresso.openapi.models import OAuthFlows as OAuthFlowsModel
-from xpresso.typing import Annotated
-
-
-class OAuth2PasswordRequestForm(BaseModel):
-    """
-    This is a dependency class, use it like:
-
-        @app.post("/login")
-        def login(form_data: Annotated[OAuth2PasswordRequestForm, Depends()]):
-            data = form_data.parse()
-            print(data.username)
-            print(data.password)
-            for scope in data.scopes:
-                print(scope)
-            if data.client_id:
-                print(data.client_id)
-            if data.client_secret:
-                print(data.client_secret)
-            return data
-
-
-    It creates the following Form request parameters in your endpoint:
-
-    grant_type: the OAuth2 spec says it is required and MUST be the fixed string "password".
-        This dependency is more leniant and allows any value.
-        To enforce this use OAuth2PasswordRequestFormStrict.
-    username: username string. The OAuth2 spec requires the exact field name "username".
-    password: password string. The OAuth2 spec requires the exact field name "password".
-    scope: Optional string. Several scopes (each one a string) separated by spaces. E.g.
-        "items:read items:write users:read profile openid"
-    client_id: optional string. OAuth2 recommends sending the client_id and client_secret (if any)
-        using HTTP Basic auth, as: client_id:client_secret
-    client_secret: optional string. OAuth2 recommends sending the client_id and client_secret (if any)
-        using HTTP Basic auth, as: client_id:client_secret
-    """
-
-    grant_type: str
-    username: str
-    password: str
-    scopes: Annotated[
-        List[str], FormEncodedField(style="spaceDelimited", explode=False)
-    ]
-    client_id: Optional[str] = None
-    client_secret: Optional[str] = None
-
-
-class OAuth2PasswordRequestFormStrict(OAuth2PasswordRequestForm):
-    grant_type: Literal["password"]
-
-
-class OAuth2(SecurityBase):
-    def __init__(
-        self,
-        *,
-        flows: Union[OAuthFlowsModel, Dict[str, Dict[str, Any]]] = OAuthFlowsModel(),
-        scheme_name: Optional[str] = None,
-        description: Optional[str] = None,
-        auto_error: Optional[bool] = True,
-    ):
-        if not isinstance(flows, OAuthFlowsModel):
-            flows = OAuthFlowsModel(**flows)
-        self.model = OAuth2Model(flows=flows, description=description)
-        self.scheme_name = scheme_name or self.__class__.__name__
-        self.auto_error = auto_error
-
-    async def __call__(self, request: Request) -> Optional[str]:  # type: ignore  # for Pylance
-        authorization: str = request.headers.get("Authorization")
-        if not authorization:
-            if self.auto_error:
-                raise HTTPException(
-                    status_code=HTTP_401_UNAUTHORIZED, detail="Not authenticated"
-                )
-            else:
-                return None
-        return authorization
-
-
-class OAuth2PasswordBearer(OAuth2):
-    def __init__(
-        self,
-        tokenUrl: str,
-        scheme_name: Optional[str] = None,
-        scopes: Optional[Dict[str, str]] = None,
-        description: Optional[str] = None,
-        auto_error: bool = True,
-    ):
-        if not scopes:
-            scopes = {}
-        flows = OAuthFlowsModel(
-            **{"password": {"tokenUrl": tokenUrl, "scopes": scopes}}
-        )
-        super().__init__(
-            flows=flows,
-            scheme_name=scheme_name,
-            description=description,
-            auto_error=auto_error,
-        )
-
-    async def __call__(self, request: Request) -> Optional[str]:
-        authorization: str = request.headers.get("Authorization")
-        scheme, param = get_authorization_scheme_param(authorization)
-        if not authorization or scheme.lower() != "bearer":
-            if self.auto_error:
-                raise HTTPException(
-                    status_code=HTTP_401_UNAUTHORIZED,
-                    detail="Not authenticated",
-                    headers={"WWW-Authenticate": "Bearer"},
-                )
-            else:
-                return None
-        return param
-
-
-class OAuth2AuthorizationCodeBearer(OAuth2):
-    def __init__(
-        self,
-        authorizationUrl: str,
-        tokenUrl: str,
-        refreshUrl: Optional[str] = None,
-        scheme_name: Optional[str] = None,
-        scopes: Optional[Dict[str, str]] = None,
-        description: Optional[str] = None,
-        auto_error: bool = True,
-    ):
-        if not scopes:
-            scopes = {}
-        flows = OAuthFlowsModel(
-            **{
-                "authorizationCode": {
-                    "authorizationUrl": authorizationUrl,
-                    "tokenUrl": tokenUrl,
-                    "refreshUrl": refreshUrl,
-                    "scopes": scopes,
-                }
-            }
-        )
-        super().__init__(
-            flows=flows,
-            scheme_name=scheme_name,
-            description=description,
-            auto_error=auto_error,
-        )
-
-    async def __call__(self, request: Request) -> Optional[str]:
-        authorization: str = request.headers.get("Authorization")
-        scheme, param = get_authorization_scheme_param(authorization)
-        if not authorization or scheme.lower() != "bearer":
-            if self.auto_error:
-                raise HTTPException(
-                    status_code=HTTP_401_UNAUTHORIZED,
-                    detail="Not authenticated",
-                    headers={"WWW-Authenticate": "Bearer"},
-                )
-            else:
-                return None  # pragma: nocover
-        return param
