--- xpresso/binders/_security/open_id_connect_url.py
@@ -1,35 +1,0 @@
-from typing import Optional
-
-from starlette.requests import Request
-from starlette.status import HTTP_401_UNAUTHORIZED
-
-from xpresso.binders.api import SecurityBase
-from xpresso.exceptions import HTTPException
-from xpresso.openapi.models import OpenIdConnect as OpenIdConnectModel
-
-
-class OpenIdConnect(SecurityBase):
-    def __init__(
-        self,
-        *,
-        openIdConnectUrl: str,
-        scheme_name: Optional[str] = None,
-        description: Optional[str] = None,
-        auto_error: bool = True
-    ):
-        self.model = OpenIdConnectModel(
-            openIdConnectUrl=openIdConnectUrl, description=description
-        )
-        self.scheme_name = scheme_name or self.__class__.__name__
-        self.auto_error = auto_error
-
-    async def __call__(self, request: Request) -> Optional[str]:
-        authorization: str = request.headers.get("Authorization")
-        if not authorization:
-            if self.auto_error:
-                raise HTTPException(
-                    status_code=HTTP_401_UNAUTHORIZED, detail="Not authenticated"
-                )
-            else:
-                return None
-        return authorization
