--- xpresso/binders/_security/utils.py
@@ -1,8 +1,0 @@
-from typing import Tuple
-
-
-def get_authorization_scheme_param(authorization_header_value: str) -> Tuple[str, str]:
-    if not authorization_header_value:
-        return "", ""
-    scheme, _, param = authorization_header_value.partition(" ")
-    return scheme, param
